requirements.txt
termcolor
python-binance
numpy
pandas
joblib